Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-openai
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
slack-sdk
python-dotenv